+++ services/src/osm.rs
@@ -6,7 +6,7 @@ use serde::{Deserialize, Serialize, Serializer};
 
 pub type NodeId = i64;
 pub type WayId = i64;
-pub type Distance = f32;
+pub type Distance = i32;
 
 pub type WayLabels = (Cycleway, Road, Salmoning);
